+++ include/game.h
@@ -10,19 +10,23 @@
 #include "celestialobject.h"
 #include "planet.h"
 #include "spaceship.h"
+#include "navball.h"
 
 class GameScene: public Scene {
 private:
+	// Game
+	sf::View* gameview;
+	double zoom;
 	DrawGroup bodies;
-	DrawGroup gui;
 	Spaceship* ship;
 	sf::Vertex* vertex;
-	CelestialObject* selectedEntity;
+	// Gui
+	sf::View* guiview;
+	DrawGroup gui;
 	sf::Text* infoText;
 	sf::Text* infoText2;
-	sf::View* gameview;
-	sf::View* guiview;
-	double zoom;
+	CelestialObject* selectedEntity;
+	Navball* navball;
 public:
 	GameScene();
 	~GameScene();