+++ src/spaceship.cpp
@@ -1,6 +1,7 @@
 #include "spaceship.h"
 #include "graphics.h"
 
+
 Spaceship::Spaceship(std::string name) : CelestialObject(10.0f, 0, name){
     this->setFillColor(sf::Color(0,0,0,0));
     //this->setFillColor(sf::Color::Yellow);
@@ -16,13 +17,13 @@ Spaceship::Spaceship(std::string name) : CelestialObject(10.0f, 0, name){
     this->lastAngle = angle;
 }
 
-void Spaceship::draw(){
+void Spaceship::draw(sf::RenderWindow& window){
     if (this->angle != this->lastAngle){
-        this->sprite.setRotation((angle/3.14)*180);
+        this->sprite.setRotation(-(angle/3.14)*180+270);
         this->lastAngle = angle;
     }
     this->sprite.setPosition(this->getPosition());
-    getWindow()->draw(this->sprite);
+    window.draw(this->sprite);
 }
 
 void Spaceship::rotate(float amount){
@@ -30,6 +31,10 @@ void Spaceship::rotate(float amount){
 }
 
 void Spaceship::throttle(float amount){
-    this->accelerate(std::sin(angle)*amount,
-                     std::cos(angle)*amount);
+    this->accelerate(-std::cos(angle)*amount,
+                     std::sin(angle)*amount);
+}
+
+float Spaceship::getAngle(){
+    return this->angle;
 }